requirements.txt
python==3.10.14

Flask==3.0.3
Flask-RESTFul==0.3.10

mysqlclient==2.2.4
SQLAlchemy==2.0.31
Flask-SQLAlchemy==3.1.1

PyJWT==2.8.0